Deduplicate OrderedFloat and merge_results, update GSPS magic
- Extract OrderedFloat and merge_top_k_results() into vector/mod.rs,
  replacing identical copies in distributed.rs and sharded.rs (~130
  lines of duplication eliminated)
- Update sparse index magic bytes from "GSPS" (legacy GallifreyDB) to
  "ASPS" (AletheiaDB SParse Search) in sparse.rs
